+++ source/isaaclab/isaaclab/sensors/pva/pva_data.py
@@ -14,12 +14,13 @@ from isaaclab.utils.backend_utils import FactoryBase
 from .base_pva_data import BasePvaData
 
 if TYPE_CHECKING:
+    from isaaclab_newton.sensors.pva import PvaData as NewtonPvaData
     from isaaclab_physx.sensors.pva import PvaData as PhysXPvaData
 
 
 class PvaData(FactoryBase, BasePvaData):
     """Factory for creating PVA data instances."""
 
-    def __new__(cls, *args, **kwargs) -> BasePvaData | PhysXPvaData:
+    def __new__(cls, *args, **kwargs) -> BasePvaData | NewtonPvaData | PhysXPvaData:
         """Create a new instance of PVA data based on the backend."""
         return super().__new__(cls, *args, **kwargs)